src/main/java/com/bosch/stocktoship/service/FeedbackService.java
package com.bosch.stocktoship.service;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.bosch.stocktoship.entity.Feedback;
import com.bosch.stocktoship.repository.FeedbackRepository;

@Service
public class FeedbackService {

	@Autowired
	private FeedbackRepository feedbackRepository;
	
	public Feedback saveFeedback(Feedback feedback) {
		return feedbackRepository.save(feedback);
	}
	
	public List<Feedback> getAllFeedbacks(){
		return feedbackRepository.findAll();
	}
}
